requirements.txt
# requirements.txt

beautifulsoup4
requests
google-cloud-storage
google-cloud-secret-manager
pdfplumber
PyPDF2
lxml
tiktoken
watchfiles
python-dotenv
fastapi
uvicorn
google-cloud-storage>=2.17.0
python-multipart
docx2txt
pymysql
cryptography
sqlalchemy
msal
